Rename flash -> flashing module
